Use Github use id to create jwt token Validate user token and let user access api TODO: Create user id in db with github id associated to it

+++ back/src/lib.rs
@@ -18,7 +18,7 @@ pub struct UserReq {
 }
 #[derive(Serialize, Deserialize, Debug)]
 pub struct User {
-    id: u32,
+    id: usize,
     name: String,
     location: String,
     email: Option<String>,
